[Refactor] MyPageResponse : controller -> service
