+++ src/main/java/com/inq/wishhair/wesharewishhair/user/service/MyPageService.java
@@ -1,6 +1,6 @@
 package com.inq.wishhair.wesharewishhair.user.service;
 
-import com.inq.wishhair.wesharewishhair.user.controller.dto.response.MyPageResponse;
+import com.inq.wishhair.wesharewishhair.user.service.dto.response.MyPageResponse;
 import com.inq.wishhair.wesharewishhair.user.domain.User;
 import com.inq.wishhair.wesharewishhair.user.domain.point.PointHistory;
 import lombok.RequiredArgsConstructor;
